Refactor and add to more models
