+++ cms/core/utils.py
@@ -1,9 +1,17 @@
+import io
 from typing import TYPE_CHECKING, Any, Optional, TypedDict
 
+import matplotlib as mpl
+import matplotlib.pyplot as plt
 from django.conf import settings
 from django.db.models import QuerySet
 from django.utils.formats import date_format
 from django.utils.translation import gettext_lazy as _
+
+FORMULA_INDICATOR = "$$"
+
+# Use LaTeX to render text in matplotlib
+mpl.rcParams.update({"text.usetex": True})
 
 if TYPE_CHECKING:
     from django.http import HttpRequest
@@ -54,3 +62,30 @@ def get_client_ip(request: "HttpRequest") -> str | None:
     if settings.IS_EXTERNAL_ENV:
         raise RuntimeError("Cannot get client IP in external environment.")
     return request.META.get("REMOTE_ADDR")
+
+
+def latex_formula_to_svg(latex: str, *, fontsize: int = 18, transparent: bool = True) -> str:
+    """Generates an SVG string from a LaTeX expression.
+
+    Args:
+        latex (str): The LaTeX string to render.
+        fontsize (int, optional): The font size for the LaTeX output. Defaults to 18.
+        transparent (bool, optional): If True, the SVG will have a transparent background. Defaults to True.
+
+    Returns:
+        str: A string containing the SVG representation of the LaTeX expression.
+    """
+    fig = plt.figure()
+    svg_buffer = io.StringIO()
+    try:
+        fig.text(0, 0, rf"${latex}$", fontsize=fontsize)
+        fig.savefig(svg_buffer, format="svg", bbox_inches="tight", transparent=transparent)
+        svg_string = svg_buffer.getvalue()
+    finally:
+        plt.close(fig)
+        svg_buffer.close()
+
+    # Remove first 3 lines of the SVG string
+    svg_string = "\n".join(svg_string.split("\n")[3:])
+
+    return svg_string
