+++ cms/standard_pages/models.py
@@ -9,6 +9,7 @@ from wagtail.search import index
 from cms.core.blocks.related import RelatedContentBlock
 from cms.core.blocks.stream_blocks import CoreStoryBlock
 from cms.core.fields import StreamField
+from cms.core.forms import PageWithEquationsAdminForm
 from cms.core.models import BasePage
 from cms.core.widgets import date_widget
 from cms.taxonomy.mixins import GenericTaxonomyMixin
@@ -20,6 +21,8 @@ if TYPE_CHECKING:
 
 class InformationPage(GenericTaxonomyMixin, BasePage):  # type: ignore[django-manager-missing]
     """A generic information page model."""
+
+    base_form_class = PageWithEquationsAdminForm
 
     template = "templates/pages/information_page.html"
